% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.narrowPeaks}
\alias{import.narrowPeaks}
\title{Import narrowPeaks file}
\usage{
import.narrowPeaks(file, extraCols = c(FoldChange = "numeric",
  MinusLog10pvalue = "numeric", MinusLog10qvalue = "numeric",
  RelativeSummitPositionToPeakStart = "numeric"), ...)
}
\arguments{
\item{file}{Path a "NAME_peaks.narrowPeak" file.}

\item{extraCols}{A character vector in the same form as colClasses from \code{\link{read.table.}}.
It should indicate the name and class of each extra/special column to read from the BED file.
See \code{\link{import.bed}} for details.}

\item{...}{Additional arguments passed to \code{import.bed}.}
}
\value{
A \code{\link{GRanges}} object.
}
\description{
From the MACS2 documentation: "NAME_peaks.narrowPeak is BED6+4 format file which contains the peak locations together with peak summit, pvalue and qvalue.
Definition of some specific columns are:
}
\details{
\itemize{
\item 5th: integer score for display calculated as int(-10*log10qvalue).
Please note that currently this value might be out of the [0-1000] range defined in UCSC Encode narrowPeak format.
\item 7th: fold-change
\item 8th: -log10pvalue
\item 9th: -log10qvalue
\item 10th: relative summit position to peak start"
}
}
\examples{
# Example file ----
file <- system.file(package = "macs2r", "extdata", "example_peaks.narrowPeak")

# Example usage ----
gr <- import.narrowPeaks(file)

# seqlevels(x) may need to be switched to the adequate format
# for downstream analyses
library(GenomeInfoDb)
seqlevelsStyle(narrowPeakGranges) <- "Ensembl"
}
\author{
Kevin Rue-Albrecht
}
