% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs2r-package.R
\docType{package}
\name{macs2r-package}
\alias{macs2r}
\alias{macs2r-package}
\title{macs2r: Helper Functions To Work With MACS2}
\description{
Provides helper functions to work with MACS2 output files.
}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

}
\keyword{internal}
